Solve my puzzle

+++ src/09-mirage-maintenance.h
@@ -5,5 +5,5 @@ using namespace std;
 class Solution_09
 {
     public:
-    int part1(vector<vector<int>> input);
+    long part1(vector<vector<long>> input);
 };
